Support readTo of Buffer in Node < 1.5.0
In Node pre-1.5.0, Buffer did not have a .indexOf method.  Add an
optional dependency on buffertools (since native modules often cause
install issues) and a dependency on buffer-indexof-polyfill as a
fallback.  Use these when Buffer.prototype.indexOf is not present.

Also update documentation to mention this limitation and fix one failing
test which relied on Buffer.prototype.indexOf.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>