Pass current chunk to until function
For efficiency, it is often preferable to deal with only the most recent
chunk of data read, to avoid repeatedly processing the same data.  Since
this is not always the case (e.g. string and RegEx matching, where
lookback buffering would be necessary), we still pass the entire buffer
as the first argument.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>