Be stricter about stream delays in specs
Writable is not specified to wait a tick before writing data.  We should
not rely on this behavior, which would cause test failure for pre-0.10
streams or custom Writable-like streams which don't follow this behavior.

Also wait a tick before calling read after .end() to ensure the 'end'
event isn't delayed a tick.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>