Add ended argument to until and endOK to readTo
In many cases callers will want to read delimiter-separated data
including the last segment which ends without a delimiter.  To do this,
add an endOK option to readTo and, more generally, call the until
function after 'end' occurs to allow it to decide whether or not to
accept the data (and how much).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>