Depend on yaku for synchronous reads
Since most promise libraries delay the call to their onFulfilled and
onRejected functions (as required by Promises/A+ 1.1 section 2.2.4),
callers may miss events which occur between resolution and the next read
call.  To avoid this, return an instance of yaku with synchronous
resolution for streams in flowing mode.

This also removes most of the value of any-promise, since we have to
depend on a specific promise library anyway.  Since callers can specify
their preferred promise class as an option, I don't see enough value to
use the any-promise module.  If users want it, they can open an issue.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>