Use async promises for argument errors
This is necessary to avoid unhandledRejection, which is emitted before
Promise.reject returns for synchronous Yaku.

An exception could be thrown instead, but this requires calling code to
add a separate error handling mechanism and complicate their code.

The unhandledRejection could be suppressed by a variety of means
(e.g. chaining a promise which discards the error before returning), but
this could make usage errors harder to detect.

Using an async promise risks missing stream events when flowing, but
this seems like the least-bad option, since it is unlikely the calling
code can recover and continue reading anyway.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>