Check for Promise.config for cancellation
There are several libraries which pass a progress function as the third
argument to the executor function.  Although I haven't seen one which
also has a cancel method, it would not surprise me if one exists now or
will exist soon.  To prevent calling a non-cancel function, also check
for Promise.config.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>