Revert "Support npm pre-1.3.7"
Since this module's dependencies use carat versioning of their
dependencies, support for old npm would require depending on older
versions of those packages which avoid the carat operator (if such
versions exist).  This is contrary to the goal of supporting the latest
package versions (for all sorts of good reasons).

This reverts commit 3f6b66f2965d25581ac24fb51f15d556e9aa9446.